Switch to passing in Double to formatter rather than NSNumber

+++ InchCalc/Domain/Value.swift
@@ -101,7 +101,7 @@ extension Value {
       return formatter.string(from: aNumber) ?? ""
 
     case let .inches(theInches):
-      return imperialFormatter(theInches)
+      return imperialFormatter(theInches.doubleValue)
     }
   }
 }
